Added paddlemovement component, refactored the Vector2D to use floats instead of ints

+++ C++Innlevering01/Vector2D.cpp
@@ -5,7 +5,7 @@ Vector2D::Vector2D()
 {
 }
 
-Vector2D::Vector2D(int x, int y)
+Vector2D::Vector2D(float x, float y)
 	: X(x), Y(y)
 {
 }
