+++ C++Innlevering01/Vector2D.h
@@ -3,13 +3,13 @@ class Vector2D
 {
 public:
 	Vector2D();
-	Vector2D(int x, int y);
+	Vector2D(float x, float y);
 	~Vector2D();
 	Vector2D operator+(const Vector2D & v) const;
 	Vector2D Vector2D::operator*(const float & s) const;
 	bool operator==(const Vector2D & v) const;
 	bool operator!=(const Vector2D & v) const;
-	int X;
-	int Y;
+	float X;
+	float Y;
 };
